AVRO-400. Adding warning for unused parameters (contributed by Eli Collins)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/avro/trunk@907153 13f79535-47bb-0310-9956-ffa450edef68

+++ lang/c/src/schema.h
@@ -61,7 +61,7 @@ struct avro_union_schema_t {
 struct avro_fixed_schema_t {
 	struct avro_obj_t obj;
 	const char *name;
-	size_t size;
+	int64_t size;
 };
 
 struct avro_link_schema_t {
